requirements.txt
streamlit
langchain-cohere
langchain
langchain-pinecone
python-dotenv
langchain_community
langchain_core
langchain-groq